[Improve][Core] Move MultiTableSink to seatunnel-api module
